Move dryRunEditor
